# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ordo Realitas RPG API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all RPG classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RpgClass"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a new class (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RpgClass"
  /classes/{id}:
    put:
      operationId: updateClass
      tags: [classes]
      summary: Update a class (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RpgClass"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /origins:
    get:
      operationId: listOrigins
      tags: [origins]
      summary: List all origins
      responses:
        "200":
          description: List of origins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Origin"
    post:
      operationId: createOrigin
      tags: [origins]
      summary: Create a new origin (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOriginBody"
      responses:
        "201":
          description: Origin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Origin"
  /origins/{id}:
    put:
      operationId: updateOrigin
      tags: [origins]
      summary: Update an origin (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOriginBody"
      responses:
        "200":
          description: Origin updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Origin"
    delete:
      operationId: deleteOrigin
      tags: [origins]
      summary: Delete an origin (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /pericias:
    get:
      operationId: listPericias
      tags: [skills]
      summary: List all pericias (skills)
      responses:
        "200":
          description: List of pericias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pericia"
    post:
      operationId: createPericia
      tags: [skills]
      summary: Create a new pericia (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePericiaBody"
      responses:
        "201":
          description: Pericia created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pericia"
  /pericias/{id}:
    put:
      operationId: updatePericia
      tags: [skills]
      summary: Update a pericia (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePericiaBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pericia"
    delete:
      operationId: deletePericia
      tags: [skills]
      summary: Delete a pericia (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /rituals:
    get:
      operationId: listRituals
      tags: [rituals]
      summary: List all rituals
      responses:
        "200":
          description: List of rituals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ritual"
    post:
      operationId: createRitual
      tags: [rituals]
      summary: Create a new ritual (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRitualBody"
      responses:
        "201":
          description: Ritual created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ritual"
  /rituals/{id}:
    put:
      operationId: updateRitual
      tags: [rituals]
      summary: Update a ritual (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRitualBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ritual"
    delete:
      operationId: deleteRitual
      tags: [rituals]
      summary: Delete a ritual (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List all items
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [items]
      summary: Create a new item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    put:
      operationId: updateItem
      tags: [items]
      summary: Update an item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/users/{id}/role:
    put:
      operationId: updateUserRole
      tags: [auth]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
components:
  schemas:
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
      required:
        - id
        - role
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
      required:
        - role
...
    RpgClass:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        descricao:
          type: string
        pvInicial:
          type: integer
        pvPorNivel:
          type: integer
        peInicial:
          type: integer
        pePorNivel:
          type: integer
        sanInicial:
          type: integer
        sanPorNivel:
          type: integer
        periciasTreindasBase:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - nome
    CreateClassBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        pvInicial:
          type: integer
        pvPorNivel:
          type: integer
        peInicial:
          type: integer
        pePorNivel:
          type: integer
        sanInicial:
          type: integer
        sanPorNivel:
          type: integer
        periciasTreindasBase:
          type: integer
      required:
        - nome
        - pvInicial
        - pvPorNivel
        - peInicial
        - pePorNivel
        - sanInicial
        - sanPorNivel
    Origin:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        descricao:
          type: string
        periciasConcedidas:
          type: array
          items:
            type: string
        poderConcedido:
          type: string
        createdAt:
          type: string
      required:
        - id
        - nome
    CreateOriginBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        periciasConcedidas:
          type: array
          items:
            type: string
        poderConcedido:
          type: string
      required:
        - nome
    Pericia:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        atributoBase:
          type: string
          enum: [FOR, AGI, INT, VIG, PRE]
        somenteTrainada:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - nome
        - atributoBase
    CreatePericiaBody:
      type: object
      properties:
        nome:
          type: string
        atributoBase:
          type: string
          enum: [FOR, AGI, INT, VIG, PRE]
        somenteTrainada:
          type: boolean
      required:
        - nome
        - atributoBase
    Ritual:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        elemento:
          type: string
          enum: [Sangue, Morte, Conhecimento, Energia, Medo]
        circulo:
          type: integer
        execucao:
          type: string
        alcance:
          type: string
        alvo:
          type: string
        duracao:
          type: string
        resistencia:
          type: string
        custoPe:
          type: integer
        descricao:
          type: string
        createdAt:
          type: string
      required:
        - id
        - nome
        - elemento
        - circulo
    CreateRitualBody:
      type: object
      properties:
        nome:
          type: string
        elemento:
          type: string
          enum: [Sangue, Morte, Conhecimento, Energia, Medo]
        circulo:
          type: integer
        execucao:
          type: string
        alcance:
          type: string
        alvo:
          type: string
        duracao:
          type: string
        resistencia:
          type: string
        custoPe:
          type: integer
        descricao:
          type: string
      required:
        - nome
        - elemento
        - circulo
    Item:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        tipo:
          type: string
          enum: [Arma, Protecao, Utilitario, Amaldicoado, Outro]
        descricao:
          type: string
        espacos:
          type: number
        categoria:
          type: string
        peso:
          type: number
        preco:
          type: integer
        requisitos:
          type: object
        createdAt:
          type: string
      required:
        - id
        - nome
        - tipo
...
    CreateItemBody:
      type: object
      properties:
        nome:
          type: string
        tipo:
          type: string
          enum: [Arma, Protecao, Utilitario, Amaldicoado, Outro]
        descricao:
          type: string
        espacos:
          type: number
        categoria:
          type: string
        peso:
          type: number
        preco:
          type: integer
        requisitos:
          type: object
      required:
        - nome
        - tipo