# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateItemBodyRequisitos } from "./createItemBodyRequisitos";
import type { CreateItemBodyTipo } from "./createItemBodyTipo";

export interface CreateItemBody {
  nome: string;
  tipo: CreateItemBodyTipo;
  descricao?: string;
  espacos?: number;
  categoria?: string;
  peso?: number;
  preco?: number;
  requisitos?: CreateItemBodyRequisitos;
}