# lib/api-zod/src/generated/types/createOriginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOriginBody {
  nome: string;
  descricao?: string;
  periciasConcedidas?: string[];
  poderConcedido?: string;
}