# lib/api-zod/src/generated/types/createPericiaBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePericiaBodyAtributoBase } from "./createPericiaBodyAtributoBase";

export interface CreatePericiaBody {
  nome: string;
  atributoBase: CreatePericiaBodyAtributoBase;
  somenteTrainada?: boolean;
}