# lib/api-zod/src/generated/types/createRitualBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRitualBodyElemento } from "./createRitualBodyElemento";

export interface CreateRitualBody {
  nome: string;
  elemento: CreateRitualBodyElemento;
  circulo: number;
  execucao?: string;
  alcance?: string;
  alvo?: string;
  duracao?: string;
  resistencia?: string;
  custoPe?: number;
  descricao?: string;
}