# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemRequisitos } from "./itemRequisitos";
import type { ItemTipo } from "./itemTipo";

export interface Item {
  id: string;
  nome: string;
  tipo: ItemTipo;
  descricao?: string;
  espacos?: number;
  categoria?: string;
  peso?: number;
  preco?: number;
  requisitos?: ItemRequisitos;
  createdAt?: string;
}