# lib/api-zod/src/generated/types/pericia.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { PericiaAtributoBase } from "./periciaAtributoBase";

export interface Pericia {
  id: string;
  nome: string;
  atributoBase: PericiaAtributoBase;
  somenteTrainada?: boolean;
  createdAt?: string;
}