# lib/api-zod/src/generated/types/ritual.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { RitualElemento } from "./ritualElemento";

export interface Ritual {
  id: string;
  nome: string;
  elemento: RitualElemento;
  circulo: number;
  execucao?: string;
  alcance?: string;
  alvo?: string;
  duracao?: string;
  resistencia?: string;
  custoPe?: number;
  descricao?: string;
  createdAt?: string;
}